requirements.txt
# UI
streamlit>=1.20.0

# LangChain ecosystem
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# Embeddings and vector store
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Database
supabase>=2.0.0

# Utilities
python-dotenv>=1.0.0
pandas>=1.5.0
pydantic>=2.0.0
typing-extensions>=4.0.0

# Document parsing
PyPDF2>=3.0.0
python-docx>=0.8.11
